requirements.txt
# using python3.10
uproot
awkward
dask-awkward
numpy
vector
pandas
correctionlib git@github.com:giorgiopizz/correctionlib
onnxruntime
hist
matplotlib
xrootd
fsspec-xrootd
requests
cloudpickle
dbs3-client
rucio-clients
numba